chore(cli): rename `3.1` folder to `openapi_3_1` ahead of AsyncAPI parsing work.
