+++ packages/cli/api-importers/openapi/openapi-v2-importer/src/extensions/x-fern-enum.ts
@@ -1,9 +1,9 @@
 import { z } from "zod";
 
-import { OpenAPIConverterContext3_1 } from "../3.1/OpenAPIConverterContext3_1";
 import { AbstractConverter } from "../AbstractConverter";
 import { AbstractExtension } from "../AbstractExtension";
 import { ErrorCollector } from "../ErrorCollector";
+import { OpenAPIConverterContext3_1 } from "../openapi_3_1/OpenAPIConverterContext3_1";
 import { FernEnumConfig } from "../types/FernEnumConfig";
 
 const CasingConfigSchema = z.object({
