+++ packages/cli/api-importers/openapi/openapi-v2-importer/src/extensions/x-fern-ignore.ts
@@ -1,7 +1,7 @@
-import { OpenAPIConverterContext3_1 } from "../3.1/OpenAPIConverterContext3_1";
 import { AbstractConverter } from "../AbstractConverter";
 import { AbstractExtension } from "../AbstractExtension";
 import { ErrorCollector } from "../ErrorCollector";
+import { OpenAPIConverterContext3_1 } from "../openapi_3_1/OpenAPIConverterContext3_1";
 
 export declare namespace FernIgnoreExtension {
     export interface Args extends AbstractConverter.Args {
