+++ packages/cli/api-importers/openapi/openapi-v2-importer/src/extensions/x-fern-server-name.ts
@@ -1,9 +1,9 @@
 import { OpenAPIV3_1 } from "openapi-types";
 
-import { OpenAPIConverterContext3_1 } from "../3.1/OpenAPIConverterContext3_1";
 import { AbstractConverter } from "../AbstractConverter";
 import { AbstractExtension } from "../AbstractExtension";
 import { ErrorCollector } from "../ErrorCollector";
+import { OpenAPIConverterContext3_1 } from "../openapi_3_1/OpenAPIConverterContext3_1";
 
 export declare namespace ServerNameExtension {
     export interface Args extends AbstractConverter.Args {
