+++ packages/cli/api-importers/openapi/openapi-v2-importer/src/index.ts
@@ -1,3 +1,3 @@
-export { OpenAPIConverter as OpenAPI3_1Converter } from "./3.1/OpenAPIConverter";
-export { OpenAPIConverterContext3_1 } from "./3.1/OpenAPIConverterContext3_1";
 export { ErrorCollector } from "./ErrorCollector";
+export { OpenAPIConverter as OpenAPI3_1Converter } from "./openapi_3_1/OpenAPIConverter";
+export { OpenAPIConverterContext3_1 } from "./openapi_3_1/OpenAPIConverterContext3_1";
