+++ packages/cli/api-importers/openapi/openapi-v2-importer/src/openapi_3_1/paths/ParameterConverter.ts
